requirements.txt
python-mermaid==0.1.3
Js2Py==0.74
python-dotenv==1.0.1
mistralai==1.5.0
boto3==1.37.6
polars==1.24.0
numpy==2.2.3
requests==2.32.3
pinecone-client=6.0.0